+++ components/opc-ua/src/Microsoft.Azure.IIoT.OpcUa.Publisher/tests/Services/PublisherJobServiceTests.cs
@@ -3,7 +3,7 @@
 //  Licensed under the MIT License (MIT). See License.txt in the repo root for license information.
 // ------------------------------------------------------------
 
-namespace Microsoft.Azure.IIoT.OpcUa.Publisher.Clients {
+namespace Microsoft.Azure.IIoT.OpcUa.Publisher.Services {
     using Microsoft.Azure.IIoT.OpcUa.Publisher.Models;
     using Microsoft.Azure.IIoT.OpcUa.Publisher.Runtime;
     using Microsoft.Azure.IIoT.OpcUa.Registry;
@@ -29,7 +29,7 @@ namespace Microsoft.Azure.IIoT.OpcUa.Publisher.Clients {
     using Xunit.Sdk;
     using Autofac;
 
-    public class PublisherJobClientTests {
+    public class PublisherJobServiceTests {
 
         [Fact]
         public async Task StartPublishTest1Async() {
